separate Minimax algorithm from PlayerMinimax
